Intersection is  complete.
This has been a useful exercise in doing basic geometry in ubik. I now know that
the API is painful to use in even contrived situations and needs a lot of
work. Rethinking it in terms of protocols for mathematical concepts somewhat
similarly to how clojure works under the hood would probably make a good start.