Reloadable app-db.
I'm still trying to rebuild re-frame from my process system. I suppose this is
useful because I want to know if it truly is general. It's also useful because
the single source of truth UI model has real advantages. It would be good to
have that at least most of the time.

Reloading is key no matter which way we go, so that effort will never be wasted.